+++ apps/chatRegistration/jest.config.ts
@@ -1,4 +1,5 @@
 /* eslint-disable */
+
 export default {
   displayName: 'chatRegistration',
   preset: '../../jest.preset.js',
@@ -7,5 +8,10 @@ export default {
     '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nrwl/react/babel'] }],
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  moduleNameMapper: {
+    '~tools/(.*)': '<rootDir>/src/tools/$1',
+    '~components/(.*)': '<rootDir>/src/components/$1',
+  },
+  setupFilesAfterEnv: ['./setupTests.ts'],
   coverageDirectory: '../../coverage/apps/chatRegistration',
 };