+++ apps/chatRegistration/src/components/App/App.component.tsx
@@ -1,10 +1,12 @@
 import { Provider } from '@teo0098/ui';
 import React from 'react';
 
+import { Registration } from '~components/Registration';
+
 export function App() {
   return (
     <Provider withGlobalStyles withNormalizeCSS>
-      <h1>Chat Registration</h1>
+      <Registration />
     </Provider>
   );
 }