+++ libs/types/src/Array/index.ts
@@ -1,0 +1,3 @@
+export type ArrayElement<T extends unknown[]> = T extends (infer R)[]
+  ? R
+  : never;