+++ libs/types/src/index.ts
@@ -1,1 +1,3 @@
+export * from './Array';
+export * from './Nullable';
 export * from './Size';