+++ libs/ui/jest.config.ts
@@ -5,6 +5,10 @@ export default {
   transform: {
     '^.+\\.[tj]sx?$': 'babel-jest',
   },
+  moduleNameMapper: {
+    '~components/(.*)': '<rootDir>/src/components/$1',
+  },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../../coverage/libs/ui',
+  setupFilesAfterEnv: ['./setupTests.ts'],
 };