+++ libs/ui/src/components/Progress/index.ts
@@ -1,0 +1,2 @@
+export * from './Progress.component';
+export * from './Progress.types';