+++ libs/ui/src/components/TextInput/index.ts
@@ -1,0 +1,2 @@
+export * from './TextInput.component';
+export * from './TextInput.types';