+++ libs/prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @teo0098/prettier
 
+## 0.0.3
+
+### Patch Changes
+
+- Change arrowParens formatting rule to always
+
 ## 0.0.2
 
 ### Patch Changes