+++ libs/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @teo0098/types
 
+## 0.1.0
+
+### Minor Changes
+
+- Add new ArrayElement type that extracts the type of an array's elements
+- Add new NotNullish type that extracts the type of a value that is not nullish
+- Add new NotNullishObj type that converts all object properties to be not nullish
+
 ## 0.0.1
 
 ### Patch Changes